python/requirements.txt
beautifulsoup4
selenium
chromedriver-binary-auto